Updated accel values to FLOAT

+++ Core/Inc/app_freertos.h
@@ -42,9 +42,9 @@ typedef struct {
   int16_t Gyro_Y;
   int16_t Gyro_Z;
 
-  int16_t Accel_X;
-  int16_t Accel_Y;
-  int16_t Accel_Z;
+  float Accel_X;
+  float Accel_Y;
+  float Accel_Z;
   
 } IMUSample_t;
 
@@ -53,9 +53,9 @@ typedef struct {
   int16_t Gyro_Y;
   int16_t Gyro_Z;
 
-  int16_t Accel_X;
-  int16_t Accel_Y;
-  int16_t Accel_Z;
+  float Accel_X;
+  float Accel_Y;
+  float Accel_Z;
 } UIState_t;
 
 /* USER CODE END ET */
